requirements.txt
pinecone>=3.0.0
openai>=1.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0

